+++ src/flask/debughelpers.py
@@ -2,9 +2,9 @@ from __future__ import annotations
 
 import typing as t
 
-from .app import Flask
 from .blueprints import Blueprint
 from .globals import request_ctx
+from .sansio.app import App
 
 
 class UnexpectedUnicodeError(AssertionError, UnicodeError):
@@ -113,7 +113,7 @@ def _dump_loader_info(loader) -> t.Generator:
         yield f"{key}: {value!r}"
 
 
-def explain_template_loading_attempts(app: Flask, template, attempts) -> None:
+def explain_template_loading_attempts(app: App, template, attempts) -> None:
     """This should help developers understand what failed"""
     info = [f"Locating template {template!r}:"]
     total_found = 0
@@ -122,7 +122,7 @@ def explain_template_loading_attempts(app: Flask, template, attempts) -> None:
         blueprint = request_ctx.request.blueprint
 
     for idx, (loader, srcobj, triple) in enumerate(attempts):
-        if isinstance(srcobj, Flask):
+        if isinstance(srcobj, App):
             src_info = f"application {srcobj.import_name!r}"
         elif isinstance(srcobj, Blueprint):
             src_info = f"blueprint {srcobj.name!r} ({srcobj.import_name})"
