+++ src/flask/logging.py
@@ -9,7 +9,7 @@ from werkzeug.local import LocalProxy
 from .globals import request
 
 if t.TYPE_CHECKING:  # pragma: no cover
-    from .app import Flask
+    from .sansio.app import App
 
 
 @LocalProxy
@@ -52,7 +52,7 @@ default_handler.setFormatter(
 )
 
 
-def create_logger(app: Flask) -> logging.Logger:
+def create_logger(app: App) -> logging.Logger:
     """Get the Flask app's logger and configure it if needed.
 
     The logger name will be the same as
